msc-5-hands/requirements.txt

# tensorflow
jupyterlab
matplotlib
enlighten ~= 1.10
scikit-learn ~= 1.0
wandb ~= 0.12
dotmap ~= 1.3
ipywidgets ~= 7.6
icecream ~= 2.1
tensorflow
tensorflow_datasets
tensorflow_probability
python-box ~= 6.0
randomname ~= 0.1
ipympl ~= 0.9
transformers
einops
pyqt5 ~= 5.15
nvsmi ~= 0.4
tensorboard_plugin_profile
holoviews[recommended]
